requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
whitenoise==6.5.0
